requirements.txt
basicsr==1.4.2
facexlib==0.3.0
insightface==0.7.3
numpy==1.26.4
onnx>=1.14.0,<=1.17.0
onnxruntime>=1.16.3,<=1.22.0
onnxruntime-gpu>=1.16.3,<=1.22.0
opencv-python>=4.7.0.72,<=4.8.1.78
Pillow>=10.0.1,<=11.3.0
pymatting<=1.1.13
spandrel==0.2.2
torch>=2.1.0,<=2.3.1
torchvision>=0.16.0,<=0.18.1
scikit-image>=0.25.0
tqdm>=4.67.1
requests>=2.32.4
setuptools~=78.1.1
nudenet==3.4.2
